update translations for queens
Update translations for the Queens release.

Note: this is being cherry-picked from stable/queens to master
instead of the other way around. The translations were actually done
based on queens, so that is where they were originally merged. In
hindsight, they should also be merged in master so that it has the
latest translations as well, so we should have followed the normal
process of merging in master first, but too late for that now.

Change-Id: I374eae9b0ef4bdd049d9b26b6bb27258c7907143
(cherry picked from commit a29ca2b75ed3a4cdd6e725f1f6a1d8e026e1771d)